wip: remove deleted_at for team

+++ crates/web/src/router/api/game/game_id/team/team_id/user/mod.rs
@@ -3,7 +3,7 @@ use cds_db::{
     entity::{team::State, user::Group},
     get_db,
 };
-use sea_orm::{ActiveModelTrait, ActiveValue::Set, ColumnTrait, EntityTrait, QueryFilter};
+use sea_orm::{ActiveModelTrait, ActiveValue::Set, ColumnTrait, EntityTrait, PaginatorTrait, QueryFilter};
 use serde::{Deserialize, Serialize};
 use serde_json::json;
 
@@ -14,11 +14,24 @@ use crate::{
 
 pub fn router() -> Router {
     Router::new()
+        // .route("/", axum::routing::get(get_member))
         .route("/", axum::routing::post(create_team_user))
         .route("/{user_id}", axum::routing::delete(delete_team_user))
         .route("/join", axum::routing::post(join_team))
         .route("/leave", axum::routing::delete(leave_team))
 }
+
+// #[derive(Clone, Debug, Serialize, Deserialize)]
+// pub struct GetMemberRequest {
+//     pub user_id: i64,
+// }
+//
+// pub async fn get_member(
+//     Extension(ext): Extension<Ext>, Path((game_id, team_id)): Path<(i64, i64)>,
+//     Json(body): Json<CreateTeamUserRequest>
+// ) -> Result<WebResponse<()>, WebError> {
+//
+// }
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct CreateTeamUserRequest {
@@ -36,6 +49,11 @@ pub async fn create_team_user(
     Json(body): Json<CreateTeamUserRequest>,
 ) -> Result<WebResponse<()>, WebError> {
     let operator = ext.operator.ok_or(WebError::Unauthorized(json!("")))?;
+
+    if operator.group != Group::Admin {
+        return Err(WebError::Forbidden(json!("")));
+    }
+
     let user = cds_db::transfer::User::from(
         cds_db::entity::user::Entity::find_by_id(body.user_id)
             .filter(cds_db::entity::user::Column::DeletedAt.is_null())
@@ -50,16 +68,13 @@ pub async fn create_team_user(
             .ok_or(WebError::BadRequest(json!("game_not_found")))?,
     );
     let team = cds_db::transfer::Team::from(
-        cds_db::entity::team::Entity::find_by_id(team_id)
-            .filter(cds_db::entity::team::Column::DeletedAt.is_null())
+        cds_db::entity::team::Entity::find()
+            .filter(cds_db::entity::team::Column::Id.eq(team_id))
+            .filter(cds_db::entity::team::Column::GameId.eq(game_id))
             .one(get_db())
             .await?
             .ok_or(WebError::BadRequest(json!("team_not_found")))?,
     );
-
-    if operator.group != Group::Admin {
-        return Err(WebError::Forbidden(json!("")));
-    }
 
     if team.state != State::Preparing {
         return Err(WebError::BadRequest(json!("team_not_preparing")));
@@ -87,18 +102,19 @@ pub async fn create_team_user(
 /// # Prerequisite
 /// - Operator is admin.
 pub async fn delete_team_user(
-    Extension(ext): Extension<Ext>, Path((_game_id, team_id, user_id)): Path<(i64, i64, i64)>,
+    Extension(ext): Extension<Ext>, Path((game_id, team_id, user_id)): Path<(i64, i64, i64)>,
 ) -> Result<WebResponse<()>, WebError> {
     let operator = ext.operator.ok_or(WebError::Unauthorized(json!("")))?;
     let team = cds_db::transfer::Team::from(
-        cds_db::entity::team::Entity::find_by_id(team_id)
-            .filter(cds_db::entity::team::Column::DeletedAt.is_null())
+        cds_db::entity::team::Entity::find()
+            .filter(cds_db::entity::team::Column::Id.eq(team_id))
+            .filter(cds_db::entity::team::Column::GameId.eq(game_id))
             .one(get_db())
             .await?
             .ok_or(WebError::BadRequest(json!("team_not_found")))?,
     );
 
-    if operator.group != Group::Admin || team.deleted_at.is_some() {
+    if operator.group != Group::Admin {
         return Err(WebError::Forbidden(json!("")));
     }
 
@@ -136,8 +152,9 @@ pub async fn join_team(
             .ok_or(WebError::BadRequest(json!("game_not_found")))?
     );
     let team = cds_db::transfer::Team::from(
-        cds_db::entity::team::Entity::find_by_id(team_id)
-            .filter(cds_db::entity::team::Column::DeletedAt.is_null())
+        cds_db::entity::team::Entity::find()
+            .filter(cds_db::entity::team::Column::Id.eq(team_id))
+            .filter(cds_db::entity::team::Column::GameId.eq(game_id))
             .one(get_db())
             .await?
             .ok_or(WebError::BadRequest(json!("team_not_found")))?,
@@ -151,7 +168,7 @@ pub async fn join_team(
         return Err(WebError::BadRequest(json!("user_already_in_game")));
     }
 
-    let criteria = cds_cache::get::<String>(format!("team:{team_id}:token"))
+    let criteria = cds_cache::get::<String>(format!("team:{team_id}:invite"))
         .await?
         .ok_or(WebError::BadRequest(json!("no_invite_token")))?;
 
@@ -173,13 +190,14 @@ pub async fn join_team(
 }
 
 pub async fn leave_team(
-    Extension(ext): Extension<Ext>, Path((_game_id, team_id)): Path<(i64, i64)>,
+    Extension(ext): Extension<Ext>, Path((game_id, team_id)): Path<(i64, i64)>,
 ) -> Result<WebResponse<()>, WebError> {
     let operator = ext.operator.ok_or(WebError::Unauthorized(json!("")))?;
 
     let team = cds_db::transfer::Team::from(
-        cds_db::entity::team::Entity::find_by_id(team_id)
-            .filter(cds_db::entity::team::Column::DeletedAt.is_null())
+        cds_db::entity::team::Entity::find()
+            .filter(cds_db::entity::team::Column::Id.eq(team_id))
+            .filter(cds_db::entity::team::Column::GameId.eq(game_id))
             .one(get_db())
             .await?
             .ok_or(WebError::BadRequest(json!("team_not_found")))?,
@@ -187,6 +205,14 @@ pub async fn leave_team(
 
     if team.state != State::Preparing {
         return Err(WebError::BadRequest(json!("team_not_preparing")));
+    }
+
+    let count = cds_db::entity::team_user::Entity::find()
+        .filter(cds_db::entity::team_user::Column::TeamId.eq(team.id))
+        .count(get_db()).await?;
+
+    if count <= 1 {
+        return Err(WebError::BadRequest(json!("team_has_no_other_member")));
     }
 
     let _ = cds_db::entity::team_user::Entity::delete_many()
